backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
mistralai>=1.0
httpx>=0.27
pydantic>=2.7
python-dotenv>=1.0
